+++ public/index.html
@@ -143,6 +143,5 @@
         <button class="load-more" type="button">load more</button>
       </section>
     </main>
-    <script src="bundle.js"></script>
-  <script type="text/javascript" src="bundle.js"></script></body>
-</html>
+    <script src="/bundle.js"></script>
+  </html>
