--- src/index.html
@@ -1,148 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
-    />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link rel="stylesheet" href="css/normalize.css" />
-    <link rel="stylesheet" href="css/style.css" />
-    <title>TASKMANAGER</title>
-  </head>
-
-  <body>
-    <div style="display: none">
-      <svg>
-        <symbol
-          id="wave"
-          viewBox="0 0 159 10"
-          preserveAspectRatio="none"
-        >
-          <g
-            fill="none"
-            fill-rule="nonzero"
-            stroke-linecap="square"
-            stroke-width="5"
-          >
-            <path
-              d="M4 4l9.257 2.463L21.367 4l7.927 2.463L38.736 4l9.117 2.463L56.103 4l8.685 2.463L73.472 4l8.684 2.463L90.84 4l8.684 2.463L108.208 4l8.684 2.463L125.576 4l8.684 2.463L142.943 4M146.423 4l9.257 2.463"
-            />
-          </g>
-        </symbol>
-      </svg>
-    </div>
-
-    <main class="main">
-      <section class="main__control control container">
-        <h1 class="control__title">UNFO. TASKMANAGER</h1>
-        <section class="control__btn-wrap">
-          <input
-            type="radio"
-            name="control"
-            id="control__task"
-            class="control__input visually-hidden"
-            checked
-          />
-          <label for="control__task" class="control__label">TASKS</label>
-          <input
-            type="radio"
-            name="control"
-            id="control__new-task"
-            class="control__input visually-hidden"
-          />
-          <label for="control__new-task" class="control__label"
-            >ADD NEW TASK</label
-          >
-          <input
-            type="radio"
-            name="control"
-            id="control__statistic"
-            class="control__input visually-hidden"
-          />
-          <label for="control__statistic" class="control__label"
-            >STATISTIC</label
-          >
-          <input
-            type="radio"
-            name="control"
-            id="control__search"
-            class="control__input visually-hidden"
-          />
-          <label for="control__search" class="control__label">SEARCH</label>
-        </section>
-      </section>
-
-      <section class="main__search search container">
-        <input
-          type="text"
-          id="search__input"
-          class="search__input search__input--hidden"
-          placeholder="START TYPING — SEARCH BY WORD, #HASHTAG OR DATE"
-        />
-        <label class="visually-hidden" for="search__input">Поиск</label>
-      </section>
-
-      <section class="main__filter filter container"></section>
-
-      <section class="statistic container visually-hidden">
-        <div class="statistic__line">
-          <div class="statistic__period">
-            <h2 class="statistic__period-title">Task Activity DIAGRAM</h2>
-
-            <div class="statistic-input-wrap">
-              <input
-                class="statistic__period-input"
-                type="text"
-                placeholder="01 Feb - 08 Feb"
-              />
-            </div>
-
-            <p class="statistic__period-result">
-              In total for the specified period
-              <span class="statistic__task-found">0</span> tasks were fulfilled.
-            </p>
-          </div>
-          <div class="statistic__line-graphic visually-hidden">
-            <canvas class="statistic__days" width="550" height="150"></canvas>
-          </div>
-        </div>
-
-        <div class="statistic__circle">
-          <div class="statistic__tags-wrap visually-hidden">
-            <canvas class="statistic__tags" width="400" height="300"></canvas>
-          </div>
-          <div class="statistic__colors-wrap visually-hidden">
-            <canvas class="statistic__colors" width="400" height="300"></canvas>
-          </div>
-        </div>
-      </section>
-
-      <section class="result container visually-hidden">
-        <button class="result__back">back</button>
-
-        <section class="result__group">
-          <h2 class="result__title">
-            #work<span class="result__count">17</span>
-          </h2>
-          <p class="result__empty">no matches found...</p>
-          <div class="result__cards"></div>
-        </section>
-      </section>
-
-      <section class="board container">
-        <p class="board__no-tasks visually-hidden">
-          Congratulations, all tasks were completed! To create a new click on
-          «add new task» button.
-        </p>
-
-        <div class="board__tasks">
-        </div>
-
-        <button class="load-more" type="button">load more</button>
-      </section>
-    </main>
-    <script src="bundle.js"></script>
-  <script type="text/javascript" src="bundle.js"></script></body>
-</html>
