+++ src/main.js
@@ -1,11 +1,8 @@
-import makeFilter from './make-filter.js';
-import makeCard from './make-card.js';
-
-const mainFilter = document.querySelector(`.main__filter`);
-const boardTasks = document.querySelector(`.board__tasks`);
+import {renderFilters} from './make-filter.js';
+import {renderCards} from './make-card.js';
+import {getRandomInteger} from './util.js';
 
 let tasksNumber = 7;
-const getRandomInteger = (min, max) => Math.round(min - 0.5 + Math.random() * (max - min + 1));
 
 const FILTERS = [
   {
@@ -40,34 +37,6 @@ const FILTERS = [
     amount: getRandomInteger(50, 150),
   },
 ];
-
-const render = (source, content) => {
-  source.innerHTML = content;
-};
-
-const renderFilters = (filters) => {
-  let content = ``;
-
-  filters.forEach((filter) => {
-    content += makeFilter(filter);
-  });
-
-  render(mainFilter, content);
-
-  mainFilter.addEventListener(`change`, () => renderCards(getRandomInteger(1, 8)));
-};
-
-const renderCards = (count) => {
-  let content = ``;
-
-  let i = 0;
-
-  while (i < count) {
-    content += makeCard(`black`, `deadline`);
-    i++;
-  }
-  render(boardTasks, content);
-};
 
 renderFilters(FILTERS);
 renderCards(tasksNumber);
