+++ src/make-filter.js
@@ -1,4 +1,10 @@
-export default (filter) => {
+import {render} from '../src/util.js';
+import {getRandomInteger} from '../src/util.js';
+import {renderCards} from './make-card.js';
+
+const mainFilter = document.querySelector(`.main__filter`);
+
+const createFilterElement = (filter) => {
   const lowerFilterName = filter.caption.toLowerCase();
   const upperFilterName = filter.caption.toUpperCase();
 
@@ -12,3 +18,17 @@ export default (filter) => {
       ${upperFilterName} <span class="filter__${lowerFilterName}-count">${filter.amount}</span>
     </label>`;
 };
+
+const renderFilters = (filters) => {
+  let content = ``;
+
+  filters.forEach((filter) => {
+    content += createFilterElement(filter);
+  });
+
+  render(mainFilter, content);
+
+  mainFilter.addEventListener(`change`, () => renderCards(getRandomInteger(1, 8)));
+};
+
+export {renderFilters};
