+++ webpack.config.js
@@ -1,5 +1,4 @@
 const path = require(`path`);
-const HtmlWebpackPlugin = require(`html-webpack-plugin`);
 
 module.exports = {
   mode: `development`,
@@ -14,11 +13,6 @@ module.exports = {
       use: `babel-loader`
     }]
   },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: `./src/index.html`
-    })
-  ],
   devtool: `source-map`,
   devServer: {
     contentBase: path.join(__dirname, `public`),
